cloud_storage: Extract materialized manifest cache tests
Move cache tests into a separate translation unit. Rename
cloud_storage_basic to clud_storage and move cache test there.
