+++ src/v/cloud_storage/tests/CMakeLists.txt
@@ -15,7 +15,7 @@ v_cc_library(
 # Regular unit tests that don't open ports and run on a single core
 rp_test(
   UNIT_TEST
-  BINARY_NAME cloud_storage_basic
+  BINARY_NAME cloud_storage
   SOURCES
     s3_imposter.cc
     directory_walker_test.cc
@@ -26,6 +26,7 @@ rp_test(
     recovery_request_test.cc
     segment_meta_cstore_test.cc
     segment_chunk_test.cc
+    materialized_manifest_cache_test.cc
   DEFINITIONS BOOST_TEST_DYN_LINK
   LIBRARIES v::seastar_testing_main Boost::unit_test_framework v::cloud_storage v::storage_test_utils v::cloud_roles v::raft
   ARGS "-- -c 1"
